Add template for git commit messages
Good commit messages help other developers understand why the code was
changed. However, it's difficult to prioritize doing this because of the
additional intentionality/work that is required. To reduce the overhead,
I'm adding a template to remind me how to write good commit messages.

As a reminder, git commit messages automatically remove comments. This
is because of the `strip` option for `commit.cleanup`.

This change was inspired by [this
gist](https://gist.github.com/lisawolderiksen/a7b99d94c92c6671181611be1641c733).